Hear me out: Modular Programming
