+++ index.html
@@ -7,7 +7,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Permissions-Policy" content="microphone=self, camera=self" />
-    <meta name="fc:frame" content='{"version":"next","imageUrl":"https://cover-art.kasra.codes/fc-audio-chat-rectangle.png","button":{"title":"Start Chatting","action":{"type":"launch_frame","name":"FC Audio Chat","url":"https://fc-audio-chat.kasra.codes","splashImageUrl":"https://cover-art.kasra.codes/fc-audio-chat-square.png","splashBackgroundColor":"#6442BF"}}}' />
+    <meta name="fc:frame" content='{"version":"next","imageUrl":"https://cover-art.kasra.codes/fc-audio-chat-rectangle.png","button":{"title":"Start Chatting","action":{"type":"launch_frame","name":"FC Audio Chat","url":"http://localhost:5173","splashImageUrl":"https://cover-art.kasra.codes/fc-audio-chat-square.png","splashBackgroundColor":"#6442BF"}}}' />
   </head>
 
   <body>
